1 - Make CORS proxy backend
CORS does not allow our frontend to call the CNB api, since the CNB api
only allows CORS from "apl.cnb.cz" (its own site).

Therefore, I made this backend that will call the CNB API (since there's
no CORS issue there), and made this endpoint callable from our frontend
(specified in the CORS_ALLOW env var).

Also, the cnb.service.ts service calls the api, but then maps it to
a standardized output.  Therefore if we were to use a different api in
the future (one that has a different response format), we would be able
to make a different plugin for that api and make it map to this
standardized interface.